requirements.txt
PasteDeploy
gunicorn
psycopg2
pyramid
pyramid_tm
sqlalchemy
zope.sqlalchemy
